chore(mjh/frontend): add ESLint config + fix setState-in-effect violations
Item 1 — ESLint config (mirrors MBK):
- Add eslint.config.js (v9 flat config) with @typescript-eslint, react-hooks,
  react-refresh; matches MBK's rule set exactly
- Install @eslint/js, typescript-eslint, eslint-plugin-react-hooks,
  eslint-plugin-react-refresh
- Downgrade react-hooks/refs + react-hooks/immutability to "warn" for 9
  pre-existing violations in useDiscoveryDefaultsPrefill.ts and
  markdown-preview.tsx (tracked in TECH_DEBT.md)
- Baseline: 11 errors before; after: 0 errors, 9 warnings

Item 2 — setState-in-effect fixes:
- DocumentEditDialog.tsx: remove useEffect sync; add key={editingDoc.id}
  at callsite in DocumentList.tsx to reset via remount
- DisplayNameSetting.tsx: split into DisplayNameForm (initialName prop,
  manages own state) + outer shell with key={currentUser?.id} — eliminates
  useEffect initialization entirely
- ResumeUploadSection.tsx: already fixed in prior PR; TECH_DEBT entry stale

lint: 0 errors | build: pass | tests: same pass/fail as pre-PR baseline
