+++ apps/myjobhunter/frontend/src/features/security/DisplayNameSetting.tsx
@@ -1,6 +1,5 @@
-import { useEffect, useState } from "react";
-import { showError, showSuccess } from "@platform/ui";
-import { useGetCurrentUserQuery, useUpdateCurrentUserMutation } from "@/lib/userApi";
+import { useGetCurrentUserQuery } from "@/lib/userApi";
+import DisplayNameForm from "@/features/security/DisplayNameForm";
 
 /**
  * Lets the user set the display name shown in their profile and on exported
@@ -8,31 +7,6 @@ import { useGetCurrentUserQuery, useUpdateCurrentUserMutation } from "@/lib/user
  */
 export default function DisplayNameSetting() {
   const { data: currentUser, isLoading } = useGetCurrentUserQuery();
-  const [updateCurrentUser, { isLoading: isSaving }] = useUpdateCurrentUserMutation();
-
-  const [name, setName] = useState("");
-  const [originalName, setOriginalName] = useState("");
-
-  useEffect(() => {
-    if (currentUser) {
-      const current = currentUser.display_name ?? "";
-      setName(current);
-      setOriginalName(current);
-    }
-  }, [currentUser]);
-
-  async function handleSave() {
-    try {
-      const trimmed = name.trim();
-      await updateCurrentUser({ display_name: trimmed || null }).unwrap();
-      setOriginalName(trimmed);
-      showSuccess("Display name saved.");
-    } catch {
-      showError("Couldn't save your name. Try again.");
-    }
-  }
-
-  const dirty = name.trim() !== originalName;
 
   return (
     <div className="space-y-2">
@@ -43,25 +17,11 @@ export default function DisplayNameSetting() {
           employers and contacts to see.
         </p>
       </div>
-      <div className="flex flex-col sm:flex-row gap-2">
-        <input
-          type="text"
-          value={name}
-          onChange={(e) => setName(e.target.value)}
-          placeholder="e.g. Jane Smith"
-          disabled={isLoading}
-          className="flex-1 px-3 py-2 text-sm border rounded-md disabled:opacity-50"
-          maxLength={100}
-        />
-        <button
-          type="button"
-          onClick={() => void handleSave()}
-          disabled={isLoading || isSaving || !dirty}
-          className="px-3 py-2 text-sm font-medium rounded-md border bg-primary text-primary-foreground hover:opacity-90 disabled:opacity-50 min-h-[44px] sm:min-h-[36px]"
-        >
-          {isSaving ? "Saving…" : "Save"}
-        </button>
-      </div>
+      <DisplayNameForm
+        key={currentUser?.id ?? "loading"}
+        initialName={currentUser?.display_name ?? ""}
+        disabled={isLoading}
+      />
     </div>
   );
 }
